refactor: enforce structural file complexity
