+++ packages/cli/src/cli/command-registry.ts
