+++ packages/cli/src/cli/path.ts
@@ -1,0 +1,5 @@
+import path from "node:path";
+
+export function relativePath(rootDir: string, filePath: string): string {
+  return path.relative(rootDir, filePath).split(path.sep).join("/");
+}
