+++ packages/cli/src/commands/check.ts
@@ -1,10 +1,11 @@
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from "node:fs";
+import { existsSync, readFileSync } from "node:fs";
 import path from "node:path";
-import { parseReviewMarkdown } from "../../application/src/index.ts";
-import { checkTaskProjection, defaultTaskProjectionPath, rebuildTaskProjection } from "../../kernel/src/index.ts";
-import { listTaskIndexPaths, readFrontmatter, readScalar, resolveHarnessLayout, taskDocumentPath } from "../../kernel/src/layout/index.ts";
-import { commandRegistry } from "./command-registry.ts";
-import type { CheckProfile, CliResult, GovernanceRebuildMode, LessonCommandMode } from "./types.ts";
+import { parseReviewMarkdown } from "../../../application/src/index.ts";
+import { checkTaskProjection } from "../../../kernel/src/index.ts";
+import { listTaskIndexPaths, readFrontmatter, readScalar, resolveHarnessLayout } from "../../../kernel/src/layout/index.ts";
+import { commandRegistry } from "../cli/command-registry.ts";
+import { relativePath } from "../cli/path.ts";
+import type { CheckProfile, CliResult } from "../cli/types.ts";
 
 interface ProfileValidationIssue {
   readonly code: string;
@@ -170,132 +171,6 @@ function validateGovernanceGeneratedViews(rootDir: string, strict: boolean): Rea
   return issues;
 }
 
-export function runGovernanceRebuild(rootDir: string, mode: GovernanceRebuildMode): CliResult {
-  const projectionPath = defaultTaskProjectionPath(rootDir);
-  const plannedRows = listTaskIndexPaths(rootDir).length;
-  if (mode === "dry-run") {
-    return {
-      ok: true,
-      command: "governance-rebuild",
-      mode,
-      rows: plannedRows,
-      projectionPath: relativePath(rootDir, projectionPath),
-      report: {
-        schema: "governance-rebuild-report/v1",
-        mode,
-        writes: [],
-        generatedViews: plannedGovernanceViews(rootDir)
-      }
-    };
-  }
-
-  const archivePath = mode === "archive" ? writeGovernanceArchive(rootDir, plannedRows) : null;
-  const result = rebuildTaskProjection({ rootDir });
-  const generated = writeGeneratedGovernanceViews(rootDir, result.rows.length);
-  return {
-    ok: true,
-    command: "governance-rebuild",
-    mode,
-    rows: result.rows.length,
-    warnings: result.warnings,
-    projectionPath: relativePath(rootDir, projectionPath),
-    generated: archivePath ? [archivePath, ...generated] : generated,
-    report: {
-      schema: "governance-rebuild-report/v1",
-      mode,
-      writes: archivePath ? [archivePath, relativePath(rootDir, projectionPath), ...generated] : [relativePath(rootDir, projectionPath), ...generated],
-      generatedViews: generated
-    }
-  };
-}
-
-export function runLessonPromote(rootDir: string, taskId: string, candidateId: string, mode: LessonCommandMode): CliResult {
-  const candidate = readLessonCandidate(rootDir, taskId, candidateId);
-  if (!candidate.ok) return candidate.result;
-  const outputPath = path.join(resolveHarnessLayout(rootDir).generatedRoot, "lessons", `${candidateId}.json`);
-  const relativeOutput = relativePath(rootDir, outputPath);
-  if (mode === "apply") {
-    mkdirSync(path.dirname(outputPath), { recursive: true });
-    writeFileSync(outputPath, JSON.stringify({
-      schema: "lesson-promotion/v1",
-      taskId,
-      candidateId,
-      title: candidate.value.title,
-      promotedAt: new Date().toISOString(),
-      source: "task-local-candidate"
-    }, null, 2), "utf8");
-  }
-  return {
-    ok: true,
-    command: "lesson-promote",
-    taskId,
-    mode,
-    generated: mode === "apply" ? [relativeOutput] : [],
-    report: {
-      schema: "lesson-promotion-report/v1",
-      mode,
-      taskId,
-      candidate: candidate.value,
-      plannedWrite: relativeOutput
-    }
-  };
-}
-
-export function runLessonSediment(rootDir: string, taskId: string, candidateId: string, title: string): CliResult {
-  const candidate = readLessonCandidate(rootDir, taskId, candidateId);
-  if (!candidate.ok) return candidate.result;
-  const outputPath = path.join(resolveHarnessLayout(rootDir).authoredRoot, "lessons", `${candidateId}.md`);
-  return {
-    ok: true,
-    command: "lesson-sediment",
-    taskId,
-    mode: "dry-run",
-    generated: [],
-    report: {
-      schema: "lesson-sediment-report/v1",
-      mode: "dry-run",
-      taskId,
-      candidate: candidate.value,
-      plannedWrite: relativePath(rootDir, outputPath),
-      title
-    }
-  };
-}
-
-function readLessonCandidate(
-  rootDir: string,
-  taskId: string,
-  candidateId: string
-): { readonly ok: true; readonly value: { readonly id: string; readonly status: string; readonly title: string } } | { readonly ok: false; readonly result: CliResult } {
-  const lessonPath = taskDocumentPath(rootDir, taskId, "lesson_candidates.md");
-  if (!existsSync(lessonPath)) {
-    return { ok: false, result: { ok: false, command: "lesson", taskId, error: { code: "lesson_candidates_missing", hint: "lesson_candidates.md is required for lesson promotion or sedimentation." } } };
-  }
-  const body = readFileSync(lessonPath, "utf8");
-  const candidate = parseLessonCandidate(body, candidateId);
-  if (!candidate) {
-    return { ok: false, result: { ok: false, command: "lesson", taskId, error: { code: "lesson_candidate_not_found", hint: `candidate not found: ${candidateId}` } } };
-  }
-  if (candidate.status !== "ready-for-review" && candidate.status !== "needs-promotion" && candidate.status !== "promoted") {
-    return { ok: false, result: { ok: false, command: "lesson", taskId, error: { code: "lesson_candidate_not_promotable", hint: `candidate ${candidateId} has status ${candidate.status}` } } };
-  }
-  return { ok: true, value: candidate };
-}
-
-function parseLessonCandidate(body: string, candidateId: string): { readonly id: string; readonly status: string; readonly title: string } | null {
-  for (const line of body.split(/\r?\n/u)) {
-    const cells = line.split("|").map((cell) => cell.trim()).filter((cell) => cell.length > 0);
-    if (cells[0] === candidateId) {
-      return {
-        id: cells[0],
-        status: cells[1] ?? "",
-        title: cells[2] ?? candidateId
-      };
-    }
-  }
-  return null;
-}
-
 function profileIssue(source: string, code: string, severity: "warning" | "hard-fail", message: string, repairHint: string): ProfileValidationIssue {
   return { source, code, severity, message, repairHint };
 }
@@ -323,43 +198,4 @@ function summarizeValidatorIssues(issues: ReadonlyArray<ProfileValidationIssue>)
 
 export function isCheckProfile(value: string): value is CheckProfile {
   return value === "source-package" || value === "private-harness" || value === "target-project";
-}
-
-function plannedGovernanceViews(rootDir: string): ReadonlyArray<string> {
-  const layout = resolveHarnessLayout(rootDir);
-  return [
-    relativePath(rootDir, defaultTaskProjectionPath(rootDir)),
-    relativePath(rootDir, path.join(layout.generatedRoot, "Harness-Ledger.md"))
-  ];
-}
-
-function writeGeneratedGovernanceViews(rootDir: string, rows: number): ReadonlyArray<string> {
-  const layout = resolveHarnessLayout(rootDir);
-  const ledgerPath = path.join(layout.generatedRoot, "Harness-Ledger.md");
-  mkdirSync(path.dirname(ledgerPath), { recursive: true });
-  writeFileSync(ledgerPath, [
-    "# Harness Ledger",
-    "",
-    "Generated projection. Authored task packages remain the source of truth.",
-    "",
-    `- Generated At: ${new Date().toISOString()}`,
-    `- Task Rows: ${rows}`,
-    ""
-  ].join("\n"), "utf8");
-  return [relativePath(rootDir, ledgerPath)];
-}
-
-function writeGovernanceArchive(rootDir: string, plannedRows: number): string {
-  const archivePath = path.join(resolveHarnessLayout(rootDir).localRoot, "archive", "governance", `${new Date().toISOString().replace(/[:.]/gu, "-")}.json`);
-  mkdirSync(path.dirname(archivePath), { recursive: true });
-  writeFileSync(archivePath, JSON.stringify({
-    schema: "governance-archive/v1",
-    archivedAt: new Date().toISOString(),
-    plannedRows
-  }, null, 2), "utf8");
-  return relativePath(rootDir, archivePath);
-}
-
-function relativePath(rootDir: string, filePath: string): string {
-  return path.relative(rootDir, filePath).split(path.sep).join("/");
 }
