+++ packages/kernel/src/projection/sqlite-task-source.ts
@@ -1,0 +1,132 @@
+import { createHash } from "node:crypto";
+import { existsSync, readdirSync, readFileSync, statSync } from "node:fs";
+import path from "node:path";
+import type { CloseoutReadiness, PackageDisposition } from "../domain/index.ts";
+import { isDomainStatus, isPackageDisposition, isTerminalStatus } from "../domain/index.ts";
+import { resolveHarnessLayout } from "../layout/index.ts";
+import type { ProjectionCanonicalStatus, CoordinationStatus, ProjectionWarning, TaskProjectionRow } from "./types.ts";
+
+export function readMarkdownSource(rootDir: string): {
+  readonly entries: ReadonlyArray<TaskSourceEntry>;
+  readonly hash: string;
+  readonly warnings: ReadonlyArray<ProjectionWarning>;
+} {
+  const tasksDir = resolveHarnessLayout(rootDir).tasksRoot;
+  if (!existsSync(tasksDir)) {
+    return { entries: [], hash: hashText("[]"), warnings: [] };
+  }
+
+  const warnings: ProjectionWarning[] = [];
+  const entries: TaskSourceEntry[] = [];
+  for (const name of readdirSync(tasksDir).sort()) {
+    const indexPath = path.join(tasksDir, name, "INDEX.md");
+    if (!existsSync(indexPath)) continue;
+    const body = readFileSync(indexPath, "utf8");
+    try {
+      entries.push({
+        taskId: name,
+        indexPath,
+        body,
+        frontmatter: parseFrontmatter(body)
+      });
+    } catch (error) {
+      warnings.push({
+        code: "source_malformed",
+        source: "source-package",
+        severity: "hard-fail",
+        message: error instanceof Error ? error.message : `Malformed task package: ${name}`,
+        repairHint: "Restore valid task-package/v2 frontmatter before running projection reads or post-merge checks."
+      });
+    }
+  }
+
+  return {
+    entries,
+    hash: hashText(JSON.stringify(entries.map((entry) => ({
+      taskId: entry.taskId,
+      sourcePath: sourcePath(rootDir, entry.indexPath),
+      body: entry.body
+    })))),
+    warnings
+  };
+}
+
+export interface TaskSourceEntry {
+  readonly taskId: string;
+  readonly indexPath: string;
+  readonly body: string;
+  readonly frontmatter: string;
+}
+
+export function taskEntryToRow(rootDir: string, entry: TaskSourceEntry): TaskProjectionRow {
+  const rawStatus = readScalar(entry.frontmatter, "  status") || "unknown";
+  const canonicalStatus = isDomainStatus(rawStatus) ? rawStatus : "unknown";
+  const rawDisposition = readScalar(entry.frontmatter, "packageDisposition") || "active";
+  const packageDisposition = isPackageDisposition(rawDisposition) ? rawDisposition : "active";
+  const lifecycleEngine = readScalar(entry.frontmatter, "  engine") || "local";
+  const source = sourcePath(rootDir, entry.indexPath);
+  return {
+    schema: "sqlite-task-row/v1",
+    taskId: readScalar(entry.frontmatter, "task_id") || entry.taskId,
+    title: readScalar(entry.frontmatter, "title") || entry.taskId,
+    canonicalStatus,
+    coordinationStatus: coordinationStatus(canonicalStatus),
+    rawStatus,
+    packageDisposition,
+    closeoutReadiness: closeoutReadiness(rootDir, entry.taskId, canonicalStatus),
+    lifecycleEngine,
+    freshness: canonicalStatus === "unknown" || !isPackageDisposition(rawDisposition) ? "stale-but-usable" : "fresh",
+    updatedAt: statSync(entry.indexPath).mtime.toISOString(),
+    source: lifecycleEngine === "local" ? "local-document" : "external-engine",
+    sourcePath: source
+  };
+}
+
+function parseFrontmatter(body: string): string {
+  const frontmatter = body.match(/^---\n([\s\S]*?)\n---/u)?.[1];
+  if (!frontmatter) throw new Error("INDEX.md missing frontmatter");
+  return frontmatter;
+}
+
+export function readScalar(frontmatter: string, key: string): string {
+  const escaped = key.replace(/[.*+?^${}()|[\]\\]/gu, "\\$&");
+  return frontmatter.match(new RegExp(`^${escaped}:[ \\t]*(.*)$`, "mu"))?.[1]?.trim() ?? "";
+}
+
+function coordinationStatus(status: ProjectionCanonicalStatus): CoordinationStatus {
+  if (status === "unknown") return "unknown";
+  if (status === "blocked") return "blocked";
+  if (status === "in_review") return "in_review";
+  return isTerminalStatus(status) ? "terminal" : "open";
+}
+
+function closeoutReadiness(rootDir: string, taskId: string, status: ProjectionCanonicalStatus): CloseoutReadiness {
+  if (status === "unknown") return "missing";
+  if (!isTerminalStatus(status) && status !== "in_review") return "not_required";
+  const taskDir = path.join(resolveHarnessLayout(rootDir).tasksRoot, taskId);
+  if (existsSync(path.join(taskDir, "walkthrough.md")) || existsSync(path.join(taskDir, "closeout.md"))) return "ready";
+  return "missing";
+}
+
+export function sourcePath(rootDir: string, filePath: string): string {
+  return path.relative(rootDir, filePath).split(path.sep).join("/");
+}
+
+export function hashExactRows(rows: ReadonlyArray<TaskProjectionRow>): string {
+  return hashText(JSON.stringify([...rows].sort(compareRows)));
+}
+
+export function hashTaskProjectionRows(rows: ReadonlyArray<TaskProjectionRow>): string {
+  return hashText(JSON.stringify([...rows].sort(compareRows).map((row) => ({
+    ...row,
+    updatedAt: "<derived-from-source-mtime>"
+  }))));
+}
+
+function hashText(text: string): string {
+  return `sha256:${createHash("sha256").update(text).digest("hex")}`;
+}
+
+export function compareRows(a: TaskProjectionRow, b: TaskProjectionRow): number {
+  return a.taskId.localeCompare(b.taskId);
+}
